% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp11.4MgDb.R
\name{rdp11.4MgDb}
\alias{rdp11.4MgDb}
\title{MgDb-class object for greengenes version 13.8 99 OTU}
\description{
rdp11.4MgDb is a MgDb-class object with
sequence and taxonomic data for the  Ribosomal Database Project Bacteria and Archaea 16S rRNA
database release 11.4, \url{https://rdp.cme.msu.edu}
}
\examples{
rdp11.4MgDb
}
\seealso{
\link[metagenomeFeatures]{MgDb-class}
}

