% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp11.5MgDb.R
\name{rdp11.5MgDb}
\alias{rdp11.5MgDb}
\title{MgDb-class object for RPD version 11.5}
\description{
rdp11.5MgDb is a MgDb-class object with sequence and taxonomic data
for the Ribosomal Database Project Bacteria and Archaea 16S rRNA
database release 11.4, \url{https://rdp.cme.msu.edu}
}
\examples{
rdp11.5MgDb
}
\seealso{
\link[metagenomeFeatures]{MgDb-class}
}
