% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribosomaldatabaseproject11.5MgDb.R
\docType{package}
\name{ribosomaldatabaseproject11.5MgDb-package}
\alias{ribosomaldatabaseproject11.5MgDb-package}
\title{RDP 16S rRNA database release 11.5 MgDb Package.}
\description{
This package loads a MgDb class object with sequence and taxonomic data for
the Ribosomal Database Project Bacteria and Archaea 16S rRNA database
release 11.5, \url{https://rdp.cme.msu.edu}
}
\examples{
ls('package:ribosomaldatabaseproject11.5MgDb')
rdp11.5MgDb
}
\seealso{
\link[metagenomeFeatures]{MgDb-class}
}
\author{
Nathan D. Olson \email{nolson@umiacs.umd.edu}
}
\keyword{package}
