% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribosomaldatabaseproject11.4MgDb.R
\docType{package}
\name{ribosonaldatabaseproject11.4MgDb-package}
\alias{ribosonaldatabaseproject11.4MgDb-package}
\title{RDP 16S rRNA database release 11.4 MgDb Package.}
\description{
This package loads a MgDb class object with
sequence and taxonomic data for the Ribosomal Database Project Bacteria and Archaea 16S rRNA
database release 11.4, \url{https://rdp.cme.msu.edu}
}
\examples{
ls('package:ribosonaldatabaseproject11.4MgDb')
rdp11.4MgDb
}
\author{
Nathan D. Olson \email{nolson@umiacs.umd.edu}
}
\seealso{
\link[metagenomeFeatures]{MgDb-class}
}
\keyword{package}

